16 sprint homework. 1. SettingActivity MVVP added
