+++ app/src/main/java/com/example/playlistmaker/App.kt
@@ -1,21 +1,29 @@
 package com.example.playlistmaker
 
 import android.app.Application
+import android.content.Context
 import android.content.res.Configuration
 import androidx.appcompat.app.AppCompatDelegate
 
 const val PLAYLIST_SHARED_PREFS = "PLAYLIST_SHARED_PREFS"
-private const val SP_THEME_KEY = "SP_THEME_KEY"
-class App : Application() {
-    var darkTheme = false
-        private set
+const val SP_THEME_KEY = "SP_THEME_KEY"
+
+class App() : Application() {
+    private var darkTheme = false
     override fun onCreate() {
         super.onCreate()
-        val sharedPrefs = getSharedPreferences(
-            PLAYLIST_SHARED_PREFS, MODE_PRIVATE)
-        darkTheme = sharedPrefs.getBoolean(SP_THEME_KEY,resources.configuration.uiMode and Configuration.UI_MODE_NIGHT_MASK ==Configuration.UI_MODE_NIGHT_YES)
+
+
+        val sharedPrefs = applicationContext.getSharedPreferences(
+            PLAYLIST_SHARED_PREFS, MODE_PRIVATE
+        )
+        darkTheme = sharedPrefs.getBoolean(
+            SP_THEME_KEY,
+            resources.configuration.uiMode and Configuration.UI_MODE_NIGHT_MASK == Configuration.UI_MODE_NIGHT_YES
+        )
         switchTheme(darkTheme)
     }
+
     fun switchTheme(darkThemeEnabled: Boolean) {
         darkTheme = darkThemeEnabled
         AppCompatDelegate.setDefaultNightMode(
@@ -26,8 +34,13 @@ class App : Application() {
             }
 
         )
-        val sharedPrefs = getSharedPreferences(
-            PLAYLIST_SHARED_PREFS, MODE_PRIVATE)
-        sharedPrefs.edit().putBoolean(SP_THEME_KEY,darkThemeEnabled).apply()
+
+
     }
+
+    fun getThemeStatus(): Boolean {
+        return darkTheme
+    }
+
+
 }
