+++ app/src/main/java/com/example/playlistmaker/Creator.kt
@@ -1,11 +1,19 @@
 package com.example.playlistmaker
 
-import com.example.playlistmaker.data.repository.PlayerGetTrackRepositoryImpl
-import com.example.playlistmaker.data.repository.mediaplayer.MediaPlayerRepositoryImpl
-import com.example.playlistmaker.domain.repository.PlayerGetTrackRepository
-import com.example.playlistmaker.domain.repository.MediaPlayerRepository
-import com.example.playlistmaker.domain.usecases.GetPlayerTrackUseCase
-import com.example.playlistmaker.domain.usecases.MediaPlayerInteractor
+import android.app.Application
+import android.content.Context
+import com.example.playlistmaker.player.data.repository.PlayerGetTrackRepositoryImpl
+import com.example.playlistmaker.player.data.repository.mediaplayer.MediaPlayerRepositoryImpl
+import com.example.playlistmaker.player.domain.repository.PlayerGetTrackRepository
+import com.example.playlistmaker.player.domain.repository.MediaPlayerRepository
+import com.example.playlistmaker.player.domain.usecases.GetPlayerTrackUseCase
+import com.example.playlistmaker.player.domain.usecases.MediaPlayerInteractor
+import com.example.playlistmaker.settings.data.repository.ThemeChangeInteractorImpl
+import com.example.playlistmaker.settings.domain.repository.OutOptionsRepository
+import com.example.playlistmaker.settings.domain.repository.ThemeChangeRepository
+import com.example.playlistmaker.settings.domain.usecases.ThemeChangeInteractor
+import com.example.playlistmaker.sharing.data.repository.OutOptionsRepositoryImpl
+import com.example.playlistmaker.sharing.domain.usecases.OutOptionsInteractor
 
 object Creator {
     fun provideGetPlayerTrackUseCase(): GetPlayerTrackUseCase {
@@ -26,5 +34,19 @@ object Creator {
         return MediaPlayerRepositoryImpl()
     }
 
+
+    fun provideThemeChangeInteractor():ThemeChangeInteractor{
+        return ThemeChangeInteractor(provideThemeChangeRepository())
+    }
+    private fun provideThemeChangeRepository(): ThemeChangeRepository {
+        return ThemeChangeInteractorImpl()
+    }
+
+    fun provideOutOptionsInteractor(application: Application):OutOptionsInteractor{
+        return OutOptionsInteractor(provideOutOptionsRepository(application))
+    }
+    private fun provideOutOptionsRepository(application: Application):OutOptionsRepository{
+        return OutOptionsRepositoryImpl(application)
+    }
 
 }
