--- app/src/main/java/com/example/playlistmaker/data/dto/TracksResponse.kt
@@ -1,6 +1,0 @@
-package com.example.playlistmaker.data.dto
-
-
-import com.example.playlistmaker.domain.models.Track
-
-class TracksResponse(val resultCounts:Int, val results:List<Track>)
