--- app/src/main/java/com/example/playlistmaker/domain/repository/MediaPlayerRepository.kt
@@ -1,12 +1,0 @@
-package com.example.playlistmaker.domain.repository
-
-
-
-interface MediaPlayerRepository  {
-    fun releasePlayer()
-    fun getStatus():Int
-    fun playTrack()
-    fun pauseTrack()
-    fun preparePlayer(dataSource:String)
-
-}
