--- app/src/main/java/com/example/playlistmaker/domain/repository/PlayerGetTrackRepository.kt
@@ -1,7 +1,0 @@
-package com.example.playlistmaker.domain.repository
-
-import com.example.playlistmaker.presentation.models.PlayerTrack
-
-interface PlayerGetTrackRepository {
-    fun getPlayerTrack(trackGson:String?): PlayerTrack
-}
