--- app/src/main/java/com/example/playlistmaker/domain/usecases/MediaPlayerInteractor.kt
@@ -1,25 +1,0 @@
-package com.example.playlistmaker.domain.usecases
-
-import com.example.playlistmaker.domain.repository.MediaPlayerRepository
-
-class MediaPlayerInteractor(
-    private val mediaPlayerRepository: MediaPlayerRepository,
-    )
- {
-    fun clickPlayTrack(){
-        mediaPlayerRepository.playTrack()
-    }
-    fun clickPauseTrack(){
-        mediaPlayerRepository.pauseTrack()
-    }
-    fun preparePlayer(dataSource:String){
-        mediaPlayerRepository.preparePlayer(dataSource)
-    }
-    fun releasePlayer(){
-        mediaPlayerRepository.releasePlayer()
-
-    }
-     fun getStatus():Int{
-         return mediaPlayerRepository.getStatus()
-     }
-}
