+++ app/src/main/java/com/example/playlistmaker/player/data/dto/TracksResponse.kt
@@ -1,0 +1,6 @@
+package com.example.playlistmaker.player.data.dto
+
+
+import com.example.playlistmaker.search.domain.models.Track
+
+class TracksResponse(val resultCounts:Int, val results:List<Track>)
