+++ app/src/main/java/com/example/playlistmaker/player/data/network/ItunesApi.kt
@@ -1,0 +1,11 @@
+package com.example.playlistmaker.player.data.network
+
+import com.example.playlistmaker.player.data.dto.TracksResponse
+import retrofit2.Call
+import retrofit2.http.GET
+import retrofit2.http.Query
+
+interface ItunesApi {
+    @GET("/search?entity=song&")
+    fun findTrack(@Query("term") term:String): Call<TracksResponse>
+}
