+++ app/src/main/java/com/example/playlistmaker/player/domain/repository/MediaPlayerRepository.kt
@@ -1,0 +1,12 @@
+package com.example.playlistmaker.player.domain.repository
+
+
+
+interface MediaPlayerRepository  {
+    fun releasePlayer()
+    fun getStatus():Int
+    fun playTrack()
+    fun pauseTrack()
+    fun preparePlayer(dataSource:String)
+
+}
