+++ app/src/main/java/com/example/playlistmaker/player/ui/models/PlayerStatus.kt
@@ -1,0 +1,8 @@
+package com.example.playlistmaker.player.ui.models
+
+enum class PlayerStatus(val status: Int) {
+    STATE_PLAYING(2),
+    STATE_PAUSED(3),
+    STATE_PREPARED(1),
+    STATE_DEFAULT(0)
+}
