+++ app/src/main/java/com/example/playlistmaker/player/ui/models/PlayerTrack.kt
@@ -1,0 +1,13 @@
+package com.example.playlistmaker.player.ui.models
+
+data class PlayerTrack(
+    val trackName: String = "",      // Название композиции
+    val artistName: String = "",     // Имя исполнителя
+    val trackTimeMillis: String = "",   // Продолжительность трека
+    val artworkUrl512: String? = "",  // Ссылка на изображение обложки
+    val collectionName: String? = "", //Альбом, может не быть
+    val releaseDate: String? = "",       //год релиза
+    val primaryGenreName: String? = "",//жанр
+    val country: String? = "",         // страна исполнителя
+    val previewUrl: String = ""
+)
