--- app/src/main/java/com/example/playlistmaker/presentation/ui/main/MainActivity.kt
@@ -1,37 +1,0 @@
-package com.example.playlistmaker.presentation.ui.main
-
-import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import android.view.View
-import android.widget.Button
-import com.example.playlistmaker.presentation.ui.media.MediaActivity
-import com.example.playlistmaker.R
-import com.example.playlistmaker.presentation.ui.search.SearchActivity
-import com.example.playlistmaker.presentation.ui.settings.SettingActivity
-
-class MainActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-
-        val searchButton = findViewById<Button>(R.id.searchButton)
-        val mediaButton = findViewById<Button>(R.id.mediaButton)
-        val settingButton = findViewById<Button>(R.id.settingsButton)
-
-        searchButton.setOnClickListener {
-            val intent = Intent(this@MainActivity, SearchActivity::class.java)
-            startActivity(intent)
-        }
-        val mediaButtonClickListener = View.OnClickListener {
-            val intent = Intent(this@MainActivity, MediaActivity::class.java)
-            startActivity(intent)
-        }
-        mediaButton.setOnClickListener(mediaButtonClickListener)
-
-        settingButton.setOnClickListener{
-            val intent = Intent(this@MainActivity, SettingActivity::class.java)
-            startActivity(intent)
-        }
-    }
-}
