+++ app/src/main/java/com/example/playlistmaker/settings/domain/repository/OutOptionsRepository.kt
@@ -1,0 +1,9 @@
+package com.example.playlistmaker.settings.domain.repository
+
+import android.content.Intent
+
+interface OutOptionsRepository {
+    fun openTerms():Intent
+    fun shareApp():Intent
+    fun writeSupport():Intent
+}
