+++ app/src/main/java/com/example/playlistmaker/settings/domain/repository/ThemeChangeRepository.kt
@@ -1,0 +1,9 @@
+package com.example.playlistmaker.settings.domain.repository
+
+import android.content.Context
+
+interface ThemeChangeRepository {
+    fun changeTheme(context:Context, checked:Boolean)
+
+    fun getTheme(context: Context):Boolean
+}
