+++ app/src/main/res/layout/activity_play_track.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.ui.player.PlayTrackActivity">
+    tools:context=".player.ui.PlayTrackActivity">
 
     <androidx.constraintlayout.widget.Guideline
         android:id="@+id/left_guideline"
