+++ app/src/main/res/layout/activity_setting.xml
@@ -14,6 +14,7 @@
             android:layout_height="wrap_content"
             android:contentDescription="@string/settActiv_backButton_descr"
             android:layout_margin="4dp"
+            android:visibility="visible"
 
             android:foreground="?attr/selectableItemBackground"
             app:layout_constraintEnd_toStartOf="@id/SettingsHeadTextBar"
